+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Rename `language/` directory to `lang/`
 - Rename `Language` setup to `Translations\Translation`
+- Split up thumbnail setup into separate setups
 
 ## 0.3.0 - 2018-08-24